requirements.txt
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2aerich==0.8.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
albucore==0.0.13
albumentations==1.4.10
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
asyncclick==8.1.8.0
asyncpg==0.30.0
attrs==25.1.0
bcrypt==4.0.1
beautifulsoup4==4.13.0
cachetools==5.5.1
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
deprecation==2.1.0
dictdiffer==0.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
fire==0.7.0
fonttools==4.55.8
frozenlist==1.5.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gotrue==2.11.3
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
iso8601==2.1.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
multidict==6.1.0
networkx==3.4.2
numpy==2.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
passlib==1.7.4
pillow==11.1.0
postgrest==0.19.3
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pypika-tortoise==0.3.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.12.1
realtime==2.3.0
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.12.0
supafunc==0.9.3
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.1.10
tomli==2.2.1
tomlkit==0.13.2
tortoise-orm==0.23.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.3
websockets==14.1
yarl==1.18.3
asttokens==3.0.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
DateTime==5.5
debugpy==1.8.14
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.0
fonttools==4.57.0
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.35.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pyzmq==26.4.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
sklearn-compat==0.1.3
spyder-kernels==3.0.3
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
xgboost==3.0.0
zope.interface==7.2
asttokens==3.0.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
DateTime==5.5
debugpy==1.8.14
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.0
fonttools==4.57.0
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.35.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
lime==0.2.0.1
llvmlite==0.44.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pyzmq==26.4.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.1
shap==0.47.2
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
spyder-kernels==3.0.3
stack-data==0.6.3
threadpoolctl==3.6.0
tifffile==2025.3.30
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
xgboost==3.0.0
zope.interface==7.2
